[global] Remaps self.Common.console
This CL changes references to self.Common.console (the global instance
of Common.Console) over to Common.Console.Console.instance(). To keep
both TypeScript and Closure happy we must make a method on the Console
class itself, since it only allows private constructors to be accessed
by static methods on the class.

Bug: 1058320
Change-Id: Icd4ac8aea76383588a5664a96643179d91812698