Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > PTAL and let me know if there is a better way to go about beforeEach and afterEach
> 
> I am a bit confused as to why they need to be separate now. What is blocking us from having a `before` and `after` in the `describe` block?
> 
> Note that you can also nest `describe` blocks. So I think your solution with `scenario` can probably be replaced by a nested `describe` instead.

The problem happens when we have multiple scenarios and need to use "beforeEach". The beforeEach block will run before each step which will cause issues with the run (even if those steps were nested inside another describe block).
The solution I was trying to find is a "beforeEach" that runs before each describe block in the scope it is in instead of before each test case.

Patch-set: 4