Fix clang-format configuration
The configuration was only including files that were in a commit diff.
This causes issues where we generate sources during the PRESUBMIT (which
are not in the git diff at that point in time) and don't format.

At the same time, we remove the complicated logic to reuse the
`.eslintignore` for clang-format purposes and use the correct
`.clang-format` "DisableFormat" options to disable formatting in various
directories.

Roll CodeMirror: ignore

Change-Id: I21b95a6e09c5d6bf036bf1aace103d6c26823ecc