Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2: Commit-Queue+1
> > 
> > > Patch Set 1:
> > > 
> > > Can you include stats on what this does to the build time?
> > 
> > It takes about 28 seconds combined for Rollup to do its work. Most of the folders take 200-300ms.
> 
> As opposed to how long without?

The 28 seconds is on top of the original build time. So the overhead of Rollup is 28 seconds, on a build that normally takes about 20-25 seconds. However, when you do a Chromium build (which is likely the only place where you will actually do this), the build is not slower because of Ninja parallelization.

Patch-set: 2
Reviewer: Gerrit User 1365394 <1365394@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>