Update patch set 4
Patch Set 4:

> 1. Change the ultimate fallback from 'en' to 'en-US' in lookupLocale(locale) and _formatIcuMessage()
> 1 - Deferring to @exterkamp. Can we do the same in LH, or will this fallback need to be configurable?

I think this makes sense. `en-US` is more specific, and could make sense here and in LH. @paulirish had more context on why `en` is an alias of `en-US` in LH, and there doesn't seem to be a good reason for it that we can find. 

In LH we would move the default locale from `en` to `en-US` and delete the `en` record in the locale.js aliases. Is that consistent with what would happen here?

Patch-set: 4
CC: Gerrit User 1312148 <1312148@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>