Don't disable clear IndexedDB object store button
Previously we disabled the clear button when there were no entries in
the table below. To be consistent with other similar views in devtools,
leave this button enabled even when there are no entries.

This also preserves tab order to make the interface more predictable
for users navigating with keyboard.

Bug: chromium:841126
Change-Id: Iccc744fd2932bc6a9a21c021c3363226bbd0f044