Make LiveLocation creation and updates asynchronous
This CL is part of the asynchronous source mapping effort.
LiveLocation update handlers need to be asynchronous in the future
as those update handlers are users of source mapping. To enable
asynchronous LiveLocation update handlers, we need to make triggering
LiveLocation updates asynchronous. This also includes LiveLocation
creation, as immediately after creation an update is scheduled.

Note that most of this CL is dedicated to housekeeping. Tests need to
be able to wait for all LiveLocation changes to settle. To this end,
we keep track of all promises that either represent a new LiveLocation
or an update and expose them to a test helper.

Test runners and web tests were already updated to use this live
location helper, but until now it was a no-op.

Bug: chromium:1032016
Change-Id: I13365c92ba667906f9836c58886abe9b9e637a6f