[styles] Make radio button focus ring circular
We can't use the outline property as this doesn't care about
border-radius which is used to make the radio button circular.

Other places in the codebase use a box shadow to fake a focus ring so
we can use that variable (0 0 0 1px var(--accent-color)) here too.

Picture of fix: https://imgur.com/Y5WXFcW

Bug: chromium:920154
Change-Id: Ie5aae884fa8e4039033141c167ce6419739e171f