[utilities] Replace Array#remove with helper in ApplicationPanelSidebar
This CL is part of a cleanup effort to get rid of Array#remove. Please
note that it would make sense to replace {_idbDatabaseTreeElements}
with a Set but two web tests access this private variable directly.
This CL replaces the single use-site with a helper call for now but
a TODO was added to convert the Array to a Set once the tests are
ported to e2e tests.

R=jacktfranklin@chromium.org

Bug: chromium:1050549
Change-Id: I89493d4561e42b8fe8291d1cef7c3b2c97ca1052