Fix input for build_release_applications
In a Rollup build (e.g. a release build with `is_debug=false`), we were
not correctly re-building when sources-that-are-rolled-up change. The
culprit is that the inputs of the build_release_applications task needs
the correct input.

Thus, always generate the input_devtools_modules, but only generate the
output_devtools_modules in the copy_devtools_modules task.

I have verified that making a change in a non-entrypoint JavaScript file
before this change caused Ninja to state "no work to do", whereas after
this change it properly rebuilds.

This should thus also fix the determinism build issue that would happen
whenever a front_end JavaScript change is rolled up into Chromium.

R=jacktfranklin@chromium.org,aerotwist@chromium.org

Bug: 1060003, 1060009, 1046596
Change-Id: I22952324ee14a1ae92070e2f524d7189e1cc9cac