Fix for flaky tests: Node re-selection
As noted in the comments, the ElementsPanel was notifying the
"selected node changed" event when the document had been cleared
out. The test wasn't accounting for this, and it doesn't actually
seem to make sense.

This change pivots the functional behavior of the panel to omit the
"selected node changed" event if there aren't any nodes in the
document. The previous behavior would fire the event, even if the
frontend didn't have a documentElement identified in the DOM.
This change pivots to skip the event in this case, because we
expect to receive DOM contents.

Bug: 1044425
Bug: 1046784
Change-Id: Ib17063d241d724b777ceda7a4c385d4784a57dab