Implement CookieReference and CookieReference revealer
Cookies are identified by a primary key (name,domain,path). This CL
adds a class `CookieReference` that represents a primary key for a
cookie, together with a associated revealer that tries to reveal one
of the cookie instances.

This doesn't always work, see associated bug, but it is a step into
the right direction.

Bug: chromium:1060505
Change-Id: Id30536ae0bfe5defb8038aa61b79f9c9600d018c