Hide experimental actions
Some actions are being loaded by the DevTools even though they depend on
code that's hidden behind an experiment. This can cause unhandled
exceptions if a user tries to execute one of them, and will pollute the
future custom keyboard shortcut list with unusable actions. To avoid
that, this CL adds experiment attributes to those actions.

Bug: 174309
Change-Id: I3da5661af95798fdaaafc08fcc6f4d393690a93f