Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Breaks clicking in a flamechart of the legacy JavaScript profiler.
> The following error is shown in devtools-on-devtools upon clicking:
> 
> Uncaught (in promise) TypeError: self.Bindings.DebuggerWorkspaceBinding.rawLocationToUILocation is not a function
>     at CPUProfileView._onEntryInvoked (ProfileView.js:38)
>     at CPUProfileFlameChart.dispatchEventToListeners (Object.js:12)
>     at CPUProfileFlameChart._onEntryInvoked (CPUProfileFlameChart.js:24)
>     at FlameChart.dispatchEventToListeners (Object.js:12)
>     at FlameChart._onClick (FlameChart.js:45)

Thanks for reporting this. While the CL seems not related to the breakage directly, the overall project (source mapping asyncification) might be. I will take a look.

Its weird though since {rawLocationToUILocation} is used in various other places in the same exact way and we haven't seen any breakage. Could be an ES Modules issue for the legacy JavaScript profiler.

Patch-set: 3