Update patch set 11
Patch Set 11: Code-Review+1

(3 comments)

LGTM for the technical implementation. I don't like the look of the logpoint icon as it doesn't fit in with the other two (shape) and features too much detail for no good reason. I've started an internal conversation about this, which I'd like to be resolved before M83, but we can go ahead with this as the technical side of the implementation looks good to me.

Patch-set: 11
Reviewer: Gerrit User 1119600 <1119600@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1