Notify user when OS theme changes, and theme is set to System Preference
If the user has their DevTools theme preferences set to
"System Preference", and then changes their OS theme when the DevTools
are opened, the new color scheme won't take effect until the user as
closed and reopened the DevTools.

With this change, the user will be presented with an InfoBar stating
that a reload is required for theme preferences to take effect.
The InfoBar will provide a button that allows the user to reload the
DevTools. If the DevTools reloaded automatically, the user would
lose any state with their current debugging sessions (e.g. pause
location, console logs, style changes, etc).

https://imgur.com/a/SHogwSU

The second part of this change is to avoid automatically reloading
the DevTools when the user changes their DevTools theme. The
user will be notified that a Reload is required, and the settings
page now has a "Reload DevTools" button so the user
can apply the changes immediately.

https://imgur.com/a/47ICsdi


Bug: 1001549
Change-Id: Ib3da1cbbddb21f1fffdeaa6519a55229e1caefc5