[global] Remaps self.Bindings.networkProjectManager
This CL changes references to self.Bindings.networkProjectManager (the
global instance of Bindings.NetworkProject.NetworkProjectManager) over
to Bindings.NetworkProject.NetworkProjectManager.instance(). To keep
both TypeScript and Closure happy we must make a method on the Settings
class itself, since it only allows private constructors to be accessed
by static methods on the class.

Bug: 1058320
Change-Id: I8adadca42dc2002d819dc42021bbaa6d80755a5a