[global] Remaps self.Bindings.resourceMapping
This CL changes references to self.Bindings.resourceMapping (the global
instance of Bindings.ResourceMapping.ResourceMapping) over to
Bindings.ResourceMapping.ResourceMapping.instance(). To keep both
TypeScript and Closure happy we must make a method on the
ResourceMapping class itself, since it only allows private constructors
to be accessed by static methods on the class.

Bug: 1058320
Change-Id: I9c90bd5aa449bf0bfdea48c7e374d8ddccc58747