Update patch set 4
Patch Set 4:

This is a lot more difficult to review than the previous CL. Mainly, I really need to see the diff from the LH version to better understand the changes.

I'm not sure at this point if the eventual goal is to modify the Lighthouse code such that CDT can consume it w/o modifications. If not - that should definitely be the goal. We (Lighthouse team) are committed to maintaining the i18n lib and these import scripts, but only if the source of truth is our repository's code - we can't commit to maintaining a fork. It's been awhile, but maybe this CL is just to show us what we need to change in LH? please correct me if I'm wrong :)

Assuming we will move forward with upstreaming these changes to LH, please tell: what are the _big_ items we need to change? We'd like to see PRs (or we could maybe do a few) of each big change coming to the LH repo.

Patch-set: 4
Reviewer: Gerrit User 1378407 <1378407@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>