Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Thanks for your cl but I think we should have some discussion around the feature.
> My MacBook for example automatically switches from light to dark them on sunset.
> Even though having that change directly reflected in DevTools would be awesome, I would
> not want it to lose my current state by reloading out of a sudden.
> 
> Mathias, Sigurd, Peter WDYT?

ya totally, initially I want to use the `applyTheme` to load the theme without reloading the devtools, sadly it doesn't work (maybe just me not that familiar with the codebase) :( so I followed the way settings is doing. But ya losing the activities is definitely not ideal (especially without a user action), that's why i also mentioned that in the commit message so we can start a discussion on this.

Patch-set: 1