Notify user when OS theme changes
If the user has their DevTools theme preferences set to
"System Preference", and then changes their OS theme when the DevTools
are opened, the new color scheme won't take effect until the user as
closed and reopened the DevTools.

With this change, the user will be presented with an InfoBar stating
that a reload is required for theme preferences to take effect.
The InfoBar will provide a button that allows the user to reload the
DevTools. If the DevTools reloaded automatically, the user would
lose any state with their current debugging sessions (e.g. pause
location, console logs, style changes, etc).

https://imgur.com/a/SHogwSU

Bug: 1001549
Change-Id: I6eb62b63823d0db1dad885d7daf436ef3173170f