Simplify list of vision deficiencies
This patch removes the UI for simulating the following vision
deficiencies:

- achromatomaly
- deuteranomaly
- protanomaly
- tritanomaly

As captured in our documentation ([1]), the abovementioned vision
deficiencies represent a broad spectrum, ranging from a slightly
reduced sensitivity to (for example) red light (in the case of
protanomaly) to a more severe lack of sensitivity to red light.
However, our built-in simulation presets can only represent a single
point on this spectrum. As Ian Hamilton points out [2], despite our
docs, this could give the false impression that most people with
protanomaly see in the way that our simulation indicates.

The most extreme form, i.e. the complete inability to perceive red
light, is protanopia instead of protanomaly. Being able to simulate
protanopia is sufficient; designing for protanopia means you’re
accessible to any severity of protanomaly anyway. The same goes for
the other vision deficiencies.

With this patch, the list of vision deficiencies DevTools can simulate
becomes:

- blurred vision
- achromatopsia
- deuteranopia
- protanopia
- tritanopia

Note that for blurred vision, the same logic applies: not everyone with
blurred vision sees in exactly the way DevTools indicates. However,
since blurred vision is a much more commonly-known vision deficiency,
the potential for confusion is lessened.

[1]: https://github.com/google/WebFundamentals/blob/c93b8e3187496ef0fa9db30bc8ccafec16025b95/src/content/en/updates/2020/03/devtools.md#emulate-vision-deficiencies--vision-deficiencies-
[2]: https://twitter.com/ianhamilton_/status/1237704178497130496

Bug: chromium:1003700
Change-Id: Icce75f0753f25dd1d641e2fffdd4519ebfdcaf04