Update patch set 2
Patch Set 2:

Oh and more thought: is it possible to write a test that actually applies one of these and verifies on the target that the colors changed? I am not sure how it is implemented, but maybe we can have an element with a background color that changes when the emulation is applied. Then we can obtain the newly computed color and assert it is not the same (or maybe it is a new color, depending on how easy it is to compute that value in the frontend)

Patch-set: 2
Reviewer: Gerrit User 1365394 <1365394@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>