Adds additional check so that style pane update is not skipped when the last property in the current styles section is deleted.
When deleting the last property in a styles section at the styles sidebar pane, the current bebavior will skip updating the pane because updatedProperty will be null due to local CSSProperty and CSSStyleDeclaration's CSSProperty being out-of-sync.
This additional check ensures that when the property's index is out-of-bound, which means it is referring to a CSSProperty that has already been deleted from its parent CSSDeclaration, it will not skip updating the pane.

Bug: chromium:1060267
Change-Id: Ie6f6a436c789745bcd1f5ece6162cc4ada82a5b3