Update patch set 5
Patch Set 5:

(2 comments)

Overall LGTM, but let’s add a TODO noting that this is a workaround for a more fundamental issue of the state going out of sync. Maybe even file a tracking bug so we can put it on the backlog?

Thanks for working on this. Very excited to see the e2e test, too!

Patch-set: 5
Reviewer: Gerrit User 1214214 <1214214@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>