Update patch set 24
Patch Set 24:

> Patch Set 24: Code-Review-1
> 
> A few thoughts:
> 
> 1. I don't expect to need to relaunch DevTools to enable.
> 2. I don't expect "Shortcuts" and "Custom shortcuts" to be in different places. That was weird.
> 3. I don't see a focus outline when I tab from the "Custom shortcuts" tab into the custom shortcuts list.

Note: While I don't expect to need to relaunch DevTools, that's not an invalid behavior, just please note it. It would be ideal if the Custom Shortcuts and Shortcuts tabs maybe were just one tab, and it toggled whether it showed the default list or the custom list based on the enabled-state of that experiment. That would solve both issues #1 and #2.

Patch-set: 24
Reviewer: Gerrit User 1329477 <1329477@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>