Update patch set 24
Patch Set 24:

(3 comments)

> Patch Set 24:
> 
> > Patch Set 24: Code-Review-1
> > 
> > A few thoughts:
> > 
> > 1. I don't expect to need to relaunch DevTools to enable.
> > 2. I don't expect "Shortcuts" and "Custom shortcuts" to be in different places. That was weird.
> > 3. I don't see a focus outline when I tab from the "Custom shortcuts" tab into the custom shortcuts list.
> 
> Note: While I don't expect to need to relaunch DevTools, that's not an invalid behavior, just please note it. It would be ideal if the Custom Shortcuts and Shortcuts tabs maybe were just one tab, and it toggled whether it showed the default list or the custom list based on the enabled-state of that experiment. That would solve both issues #1 and #2.

Added a note to the experiment description so people will know to reload. I've also added an order property to the new settings tab so that it will appear at the end of the list, like Shortcuts. I'm hesitant to actually combine them in case doing so complicates removing Shortcuts entirely later. There's also general weirdness with that tab (like localizing its id) that makes me hesitant to touch it.

The lack of keyboard navigation is a known issue and we have planned work to fix it, but that will require a lot of ListWidget changes in a future CL.

Patch-set: 24