Update patch set 5
Patch Set 5:

I disagree with @Songtao here -- we shouldn't be introducing unnecessary risk by renaming this function as part of an unrelated bugfix. If we're doing feature work, I'm okay with making changes like this; or even if we were changing the signature of the function, I'd be more okay with that. But when we're changing the internal guts of a function to turn an error case into a success case, that doesn't justify the name change.

Patch-set: 5
Reviewer: Gerrit User 1329477 <1329477@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>