Update patch set 6
Patch Set 6: Code-Review+1

This is good, it catches an error that Closure should have caught. We can definitely review this (and maybe we should log a bug, because this function is not tested in the frontend tests or Blink web tests).

Patch-set: 6
Reviewer: Gerrit User 1329477 <1329477@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1