Check Ninja output with expected build output
We now check that the build_output.txt is equivalent to the build_output.txt.expected.
This includes compilation failures, which `compilation_failure_front_end` shows.

To be able to have machine-agnostic buildoutputs, the toolchains need to be
absolutely referenced and thus every fixture needs the same copy of a toolchain.
Also, the ts_library print must use the relative path rather than the absolute path.

R=jacktfranklin@chromium.org

Bug: 1064287
Change-Id: I4a4db566cc155b7f6e73cadc2efcfa8d615286ed