Fix maybeLinkifyScriptLocation fallback strategy
Currently, there's a regression where some links in the performance
panel fail to update when the scripts that they point to are
pretty-printed. This was introduced when Components.Linkifier was made
accessible here [1] and a complicated ternary statement was rewritten
into an if-else. Dropping the else restores the previous functionality
and solves the regression. While debugging, I came across another issue
where call frame column numbers were not being used, so I fixed that as
well.

Bug: 1061379
Change-Id: I839c2db7bebadf9c8183f517554dfd5af293a349