[issues] Generate COEP releated issues in the front-end
As back-end changes are being blocked, we opted to implement issue
generation for COEP in the front-end only for now, and will more it
to the back-end at a later point in time.

This CL adds an issue generator in the front-end that scans network
activity and identifies COEP issues, filing issues for each of them.

Change-Id: I1d9baed1bc996efb94bcf0bab5cddde8dbbbe904