Enable no-case-declarations lint rule
Lexical declarations in `case` and `default` clauses are a footgun:
they are visible in the entire `switch` block, but they only get
initialized upon assignment, which only happens if the relevant
`case` is actually reached.

To ensure that such lexical declarations only apply to the current
`case` (which is usually the intention), `case` clauses containing
them should be wrapped in curly braces to create an explicit block.

Change-Id: I63d9341fcd76d4b9ce8281bd0e6573b886577f08