[ui] Add a border to text input on Workspace settings tab
The border was only shown on hover, which on a white background is
confusing as there is no indication it's a text input.

Normally these inputs are on a grey background so the border isn't
needed until hover.

Before: https://imgur.com/xuN2Q2J
After: https://imgur.com/RC35CLI

Bug: chromium:1060513
Change-Id: Ic21652f652710042d3a7611a70b5d0837383f816