Network: Fix blue focus ring on Network Log Area when Learn More Clicked
This CL fixes an issue that causes the network Log area to receive
a focus indicator with the "Learn More" link is focused/clicked.

This issue was caused because the logic to trigger the focus ring
on the network log area assumed the only focusable content within the
container was the DataGrid. It previously showed the ring whenever
focus was within the container and the datagrid did not have a request
selected.

Now it only shows if the datagrid receives focus and does not have a node
selected.

(Note that this focus logic was included because the
waterfall column is not part of the Datagrid)

Before: https://imgur.com/2cGS207
After: https://imgur.com/LzEDWBW

Bug: 1056138
Change-Id: I5c1f5fd03b5ed674f27c5b591fc76cfa4d46d165