Allow .d.ts files to be included in ts_library gn.
For ts_library build definitions, it's presently considered valid to
specify .d.ts files (there is a note about them in the BUILD.gn
definition and how they don't produce output files), but for those
files which are not part of the base set of .d.ts files included in
the BUILD.gn file, they're ignored. The build script here
unconditionally emits an expected .d.js file as part of the expected
build output (which of course doesn't exist).

This change checks .ts files to determine whether they're .d.ts
files and, if they match, skips them when calculating the emitted
files.

Change-Id: Ic72e1100eed683732ff80c57042e624914dc4b26