Update patch set 1
Patch Set 1:

Hey Tim -

Edge has some code in Root which depends on the InspectorHost and Common modules (but checks for their presence first). This wasn't a problem in the non-ESM and non-TypeScript world because those legacy modules still existed in the global namespace; but after Typescriptifying Root, this became difficult because there was no way to annotate that the existence of those and no way to create a circular dependency between Root and Common, for example.

The way this ends up being used is, for example, in root/BUILD.gn:

    ts_library("root") {
      sources = [
        "legacy-imports.edge.d.ts",
        "Runtime.js",
        "root.js",
      ]
    }


PTAL and let me know if you think this is okay. I've already merged this change into Edge, so strictly speaking it's not required in Chromium; but I think it'll be goodness nonetheless.

Thanks! -Rob

Patch-set: 1
Reviewer: Gerrit User 1182054 <1182054@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1365394 <1365394@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false