[issues] Add audits capability
Currently, workers don't support the audits domain. A WIP CL for adding
the audits domain is here: https://crrev.com/c/2119539

Until we support the audits domain, we must only enable the audits
domain for targets that support it, i.e. frames.

This CL adds a new capability Audits that only frames have, and
associates the IssuesModel with that capability, effectively ensuring
that only frames get an associated IssuesModel.


Change-Id: Ib6bc3374874d48a50f55ceb74a5af8e5ecd0a91a