Add test to check for non-determinism in tsbuildinfo
The test will check the generated .tsbuildinfo against the expected.tsbuildinfo
and ensures that the content is equal.

I initially tried to generate twice and compare the outputs, but for some reason
the two runs would always be identical (even though sometimes the content of
both files could differ between other runs). Therefore, I opted for checking
for the full output instead. Thus we will need to update the expectation file
whenever we are updating the typescript version we are using.

R=jacktfranklin@chromium.org

Bug: 1064287, 1058578, 1054494
Change-Id: I198e1907f674d930696744c18f11bf2321a8fd37