Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> `.d.ts` should not be outputted at all by TypeScript. Instead, they should be added to `GLOBAL_TYPESCRIPT_DEFINITION_FILES` in the `ts_library`.
> 
> Secondly, we already have a `front_end/legacy/legacy-defs.d.ts` so I recommend adding it to that file instead.
> 
> Lastly, if we would make this change, you have to update the `typescript.gni` logic as well to make sure that the files are correct in the `outputs` array. (See the logic around `output_files`) I will update the tests to make sure we don't break that behavior.

GLOBAL_TYPESCRIPT_DEFINITION_FILES is a Python variable. I'd prefer to be able to locally add these to a local BUILD.gn if possible; it allows the module to be self-contained.

It appears to work as expected if I add the ephemeral definitions to legacy/legacy-defs.d.ts; but I'd prefer to keep the definitions separate if possible. We like that it allows us to avoid merge conflicts except with the BUILD.gn file (but you'll note that I added .edge.d.ts to the file name so it's clear that we shouldn't ever run into a filename collision).

I'll investigate a little more what you mean about the last. I'm not sure. I don't see any new output being generated by the change (in fact, that's precisely what this change is intended to fix).

Patch-set: 1